Initial commit for moving config command
